Update README.md
The Indian Startup Ecosystem is the second largest in the world and this is a feat which did not happen by mere chance rather through planning, innovation, and supportive government interventions through the Startup Indian Initiative. The Ecosystem is made up of businesses from various sectors, in different locations, funding stages and intricate profile of the companies involved.
The period between 2018 and 2021 was particularly fascinating for the Indian startup scene. This era witnessed a confluence of factors that shaped the landscape, including regulatory changes, technological advancements, economic shifts, and societal dynamics. Startups across various sectors, including technology, e-commerce, healthcare, fintech, and more, made significant strides during this time. 
Understanding the key parameters that enabled these ventures to thrive is essential for those seeking to embark on their entrepreneurial journey within this ecosystem.
 For a new business looking to penetrate the India startup ecosystem there is the need to investigate the existing systems with the help of available data and know how feasible and appropriate it will be in other for the business to thrive, this is what the project seeks to know and achieve.